RFC: Give MethodMatch its own type
Every time I see any code dealing with the svecs we get back from
the method match code, I think that it shouldn't really be an svec:
1. It's always the same length
2. Basically every use of it typeasserts the field type
3. Every use of it needs the same magic numbers to access the fields.

All put together this should just be a type. This basically does that,
but there's a few things missing, so before I go finish it, I figured,
I'd make sure that there isn't some reason not to do this.

TODO:
1. (deprecated) getindex/iterate for `MethodMatch` to avoid breaking packages that use the internal API
2. Make the bitfield a uint8_t for compatibility with compiler that use a different layout
3. Double check the gf.c changes. I think I fatfingered one of the mappings.

Doesn't need extensive review in this state, since it isn't done,
just a quick yay/nay on the direction.